pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dns-smtp-email-validator"
version = "0.1.11"
description = """
DNS and SMTP Email Validator is a robust Python library designed
to validate email addresses at multiple levels. It checks the format, domain,
and MX records, and even communicates with the mail server to confirm the
existence of the recipient's email address.
"""
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "shailesh pandit", email = "shaileshpandit141@gmail.com" }]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "dnspython>=2.3.0",
    "load-dotenv>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/shaileshpandit141/dns-smtp-email-validator"
Repository = "https://github.com/shaileshpandit141/dns-smtp-email-validator"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
